pyproject.toml
[project]
name = "mini-timelapse"
version = "0.1.0"
description = "Compile timestamped images into timelapse videos with embedded metadata, and extract them back."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Asger Svenning" }
]
keywords = ["timelapse", "video", "exif", "metadata", "sftp"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "av>=17.0.0",
    "numpy>=2.4.4",
    "piexif>=1.1.3",
    "pillow>=12.2.0",
    "tqdm>=4.67.3",
]

[project.optional-dependencies]
remote = [
    "pyremotedata>=0.1.0",
]
test = [
    "pytest>=8.0.0",
]

[project.scripts]
timelapse-compile = "mini_timelapse.compile:main"
timelapse-decompile = "mini_timelapse.decompile:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mini_timelapse"]
